optimize bean performance
