+++ commons/gridgo-bean/src/main/java/io/gridgo/bean/serialization/msgpack/MsgpackSerializer.java
@@ -2,8 +2,6 @@ package io.gridgo.bean.serialization.msgpack;
 
 import static io.gridgo.utils.pojo.PojoUtils.getGetterProxy;
 
-import java.io.ByteArrayInputStream;
-import java.io.ByteArrayOutputStream;
 import java.io.IOException;
 import java.io.InputStream;
 import java.io.OutputStream;
@@ -12,11 +10,8 @@ import java.util.Map;
 import java.util.Map.Entry;
 
 import org.msgpack.core.MessageFormat;
-import org.msgpack.core.MessagePack;
 import org.msgpack.core.MessagePacker;
 import org.msgpack.core.MessageUnpacker;
-import org.msgpack.core.buffer.InputStreamBufferInput;
-import org.msgpack.core.buffer.OutputStreamBufferOutput;
 
 import io.gridgo.bean.BArray;
 import io.gridgo.bean.BElement;
@@ -35,42 +30,8 @@ import lombok.NonNull;
 public class MsgpackSerializer extends AbstractBSerializer {
 
     public static final String NAME = "msgpack";
-
-    private static class PackerAndBuffer implements AutoCloseable {
-        private final OutputStreamBufferOutput output = new OutputStreamBufferOutput(new ByteArrayOutputStream(0));
-        private final MessagePacker packer = MessagePack.newDefaultPacker(this.output);
-
-        @Override
-        public void close() throws IOException {
-            this.packer.close();
-            this.output.close();
-        }
-
-        public MessagePacker reset(OutputStream out) throws IOException {
-            this.output.reset(out);
-            return this.packer;
-        }
-    }
-
-    private static class UnpackerAndBuffer implements AutoCloseable {
-        private final InputStreamBufferInput input = new InputStreamBufferInput(new ByteArrayInputStream(new byte[0]));
-        private final MessageUnpacker unpacker = MessagePack.newDefaultUnpacker(this.input);
-
-        @Override
-        public void close() throws IOException {
-            this.unpacker.close();
-            this.input.close();
-        }
-
-        public MessageUnpacker reset(InputStream in) throws IOException {
-            this.input.reset(in);
-            return this.unpacker;
-        }
-    }
-
-    private final ThreadLocal<UnpackerAndBuffer> UNPACKERS = ThreadLocal.withInitial(UnpackerAndBuffer::new);
-
     private final ThreadLocal<PackerAndBuffer> PACKERS = ThreadLocal.withInitial(PackerAndBuffer::new);
+    private final ThreadLocal<UnpackerAndBuffer> UNPACKERS = ThreadLocal.withInitial(UnpackerAndBuffer::new);
 
     private void packAny(Object obj, MessagePacker packer) throws IOException {
         if (obj == null) {
@@ -78,7 +39,7 @@ public class MsgpackSerializer extends AbstractBSerializer {
             return;
         }
 
-        if (obj instanceof BElement) {
+        if (BElement.class.isInstance(obj)) {
             BElement element = (BElement) obj;
 
             if (element.isValue()) {
